Add AArch64 FEAT_SHA3 implementation of AEGIS128L
Add an implementation of AEGIS128L using the Neon FEAT_SHA3 extension.
This allows use of the EOR3 and BCAX instructions.

Running the benchmarks on Arm Neoverse micro-architectures with LLVM 20,
this gives the following changes (higher is better, 100% = no change):

                Neoverse V1     Neoverse V2     Neoverse N2
AEGIS-128L      112.62%         114.91%         116.28%
AEGIS-128L MAC  81.10%          80.32%          93.89%

The EOR3 instruction as present in the above cores has worse throughput
than the standard EOR instruction, with only one EOR3 instruction being
issued per cycle. This is compared to four EOR instructions per cycle on
Neoverse V1 and Neoverse V2, and two per cycle on Neoverse N2.

This difference in performance observation can therefore be explained as
such:

* For AEGIS-128L, the presence of a small number of EOR3 instructions
  reduces the overall vector instruction count and therefore improves
  performance. The reduced throughput of EOR3 instructions is not an
  issue here since the limiting factor is the number of other
  instructions executing on the remaining vector units.

* For AEGIS-128L MAC, there are more EOR3 instructions automatically
  emitted by the compiler combining EOR intrinsics, leading to the
  limited throughput of these instructions becoming more of an issue and
  an overall slowdown in performance.

The imbalance between EOR and EOR3 instruction throughputs is resolved
from Neoverse V3 and Neoverse N3, so it is expected that the MAC
performance inversion is limited to this current generation of
micro-architectures.

For comparison, when running on client cores which do not have the EOR3
throughput limitation, we instead observe the following improvements:

                Cortex-A520     Cortex-A720     Cortex-X4
AEGIS-128L      121.57%         117.18%         116.32%
AEGIS-128L MAC  118.90%         112.52%         112.01%