cpu.c: Add feature detection for AArch64 FEAT_SVE_AES
The FEAT_SVE_AES extension provides support for the Arm Scalable Vector
Extension (SVE) variants of the existing Neon AESE/AESMC instructions.

Additionally, the FEAT_SVE2 extension provides support for SVE variants
of the EOR3 and BCAX instructions, for which Neon support is present
under the FEAT_SHA3 extension previously added.

FEAT_SVE_AES is an optional extension introduced as part of the Arm
v9.0-A architecture. If FEAT_SVE_AES is present then FEAT_SVE2 must also
be present, so it is sufficient to only check for the presence of
FEAT_SVE_AES.

This commit adds detection logic for Windows, Apple Silicon, and
getauxval-based platforms. Since FEAT_SVE_AES also implies that
FEAT_SHA3 is present, additionally amend the FEAT_SHA3 detection to use
this in lieu of a more precise way of detecting support on Windows.